+++ physicsnemo/distributed/shard_utils/__init__.py
@@ -14,11 +14,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from physicsnemo.utils.version_check import check_module_requirements
 
-def register_shard_wrappers():
+# Prevent importing this module if the minimum version of pytorch is not met.
+try:
+    check_module_requirements("physicsnemo.distributed.shard_tensor")
 
-    from .attention_patches import sdpa_wrapper
-    from .conv_patches import generic_conv_nd_wrapper
-    from .natten_patches import na2d_wrapper
-    from .normalization_patches import group_norm_wrapper
-    from .pooling_patches import generic_avg_pool_nd_wrapper
+    def register_shard_wrappers():
+        from .attention_patches import sdpa_wrapper
+        from .conv_patches import generic_conv_nd_wrapper
+        from .natten_patches import na2d_wrapper
+        from .normalization_patches import group_norm_wrapper
+        from .pooling_patches import generic_avg_pool_nd_wrapper
+        
+except ImportError:
+    pass
+