+++ physicsnemo/distributed/shard_utils/natten_patches.py
@@ -15,19 +15,25 @@
 # limitations under the License.
 
 import importlib.util
-from typing import Any, Callable, List, Tuple, Union
+from typing import Any, Tuple, Union
 
 import torch
 import wrapt
-from torch.distributed.tensor.placement_types import Shard
 
-from physicsnemo.distributed import ShardTensor
-from physicsnemo.distributed.shard_utils.halo import (
+from physicsnemo.utils.version_check import check_module_requirements
+
+check_module_requirements("physicsnemo.distributed.shard_tensor")
+
+from torch.distributed.tensor.placement_types import Shard  # noqa: E402
+
+from physicsnemo.distributed.shard_utils.halo import ( # noqa: E402
     HaloConfig,
     halo_padding,
     unhalo_padding,
 )
-from physicsnemo.distributed.shard_utils.patch_core import (
+
+from physicsnemo.distributed import ShardTensor  # noqa: E402
+from physicsnemo.distributed.shard_utils.patch_core import (  # noqa: E402
     MissingShardPatch,
     UndeterminedShardingError,
 )