+++ physicsnemo/registry/model_registry.py
@@ -14,13 +14,24 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import warnings
 from importlib.metadata import EntryPoint, entry_points
 from typing import List, Union
 
-# This import is required for compatibility with doctests.
-import importlib_metadata
+# NOTE: This is for backport compatibility, some entry points seem to be using this old class
+# Exact cause of this is unknown but it seems to be related to multiple versions
+# of importlib being present in the environment
+ENTRY_POINT_CLASSES = [
+    EntryPoint,
+]
+try:
+    from importlib_metadata import EntryPoint as EntryPointOld  # noqa: E402
 
-import physicsnemo
+    ENTRY_POINT_CLASSES.append(EntryPointOld)
+except ImportError:
+    pass
+
+import physicsnemo  # noqa: E402
 
 
 # This model registry follows conventions similar to fsspec,
@@ -43,6 +54,22 @@ class ModelRegistry:
         entrypoints = entry_points(group="physicsnemo.models")
         for entry_point in entrypoints:
             registry[entry_point.name] = entry_point
+
+        # Pull in any modulus models for backwards compatibility
+        entrypoints = entry_points(group="modulus.models")
+        for entry_point in entrypoints:
+            if entry_point.name not in registry:
+                # Add depricated warning
+                warnings.warn(
+                    f"Model {entry_point.name} is being loaded from the 'modulus.models' group. "
+                    f"This probably means it is being exposed from a package that has not yet been "
+                    f"updated to use the 'physicsnemo.models' group. This group may be removed in a "
+                    f"future release. Please contact the package maintainer to update the entry point.",
+                    DeprecationWarning,
+                    stacklevel=2,
+                )
+                registry[entry_point.name] = entry_point
+
         return registry
 
     def register(
@@ -106,7 +133,7 @@ class ModelRegistry:
 
         model = self._model_registry.get(name)
         if model is not None:
-            if isinstance(model, (EntryPoint, importlib_metadata.EntryPoint)):
+            if isinstance(model, tuple(ENTRY_POINT_CLASSES)):
                 model = model.load()
             return model