+++ test/distributed/test_shard_tensor_redistribute.py
@@ -15,16 +15,32 @@
 # limitations under the License.
 
 import pytest
+
+from physicsnemo.utils.version_check import check_module_requirements
+
+try:
+    check_module_requirements("physicsnemo.distributed.shard_tensor")
+    from test_shard_tensor_initialization import (
+        init_dist,
+        init_global_shape_and_placements,
+    )
+    from torch.distributed.tensor.placement_types import Replicate, Shard
+
+    from physicsnemo.distributed import ShardTensor
+
+
+except ImportError:
+    pytest.skip(
+        "Skipping test because physicsnemo.distributed.shard_tensor is not available",
+        allow_module_level=True,
+    )
+
+
 import torch
 import torch.distributed as dist
 from distributed_utils_for_testing import modify_environment
-from test_shard_tensor_initialization import (
-    init_dist,
-    init_global_shape_and_placements,
-)
-from torch.distributed.tensor.placement_types import Replicate, Shard
 
-from physicsnemo.distributed import DistributedManager, ShardTensor
+from physicsnemo.distributed import DistributedManager
 
 
 def shard_tensor_factory(mesh_names, mesh_sizes, requires_grad=False, uneven=True):
@@ -135,8 +151,8 @@ def test_shard_tensor_reduction(data_parallel_size, domain_H, domain_W, op):
     if domain_H == 1 and domain_W == 1:
         pytest.skip("No point testing this without parallelism in the domain axes")
 
-    if op == torch.mean:
-        pytest.xfail("Mean reduction not yet supported for uneven tensor shapes")
+    # if op == torch.mean:
+    # pytest.xfail("Mean reduction not yet supported for uneven tensor shapes")
 
     remaining_gpus = num_gpus
     mesh_names = ["data_parallel"]
@@ -152,7 +168,7 @@ def test_shard_tensor_reduction(data_parallel_size, domain_H, domain_W, op):
         mesh_sizes.append(domain_W)
         remaining_gpus = int(remaining_gpus / domain_W)
 
-    verbose = False
+    verbose = False  # Change to True for debug
 
     torch.multiprocessing.set_start_method("spawn", force=True)
 
@@ -254,7 +270,7 @@ def test_shard_tensor_redistribute1d(data_parallel_size, domain_H, redistributio
     if case_name == "shard_to_shard" and len(mesh_sizes) <= 2:
         pytest.skip("Not enough dimensions for shard-to-shard test")
 
-    verbose = False
+    verbose = False  # Change to True for debug
 
     torch.multiprocessing.set_start_method("spawn", force=True)
 
@@ -337,7 +353,7 @@ def test_shard_tensor_redistribute2d(
     if case_name == "shard_to_shard" and len(mesh_sizes) <= 2:
         pytest.skip("Not enough dimensions for shard-to-shard test")
 
-    verbose = False
+    verbose = False  # Change to True for debug
 
     torch.multiprocessing.set_start_method("spawn", force=True)